refactor: formFields 객체 상수로 분리
